+++ src/libres/db/models/base.py
@@ -1,5 +1,11 @@
 from __future__ import annotations
 
-from sqlalchemy.ext import declarative
 
-ORMBase = declarative.declarative_base()
+from typing import TYPE_CHECKING
+if TYPE_CHECKING:
+    from sqlalchemy.ext.declarative import declarative_base
+else:
+    # FIXME: Move this import out when dropping 1.4 support
+    from sqlalchemy.orm import declarative_base
+
+ORMBase = declarative_base()